Imported the ML Agents Toolkit
Along with this I created a virtual environment for python that I ignore in the gitignore file.
When I first imported ml agents I got an error, after closing and reopening unity the error was gone. I kept a record of the error in case it breaks anything in the future.